Removed slate, more tidying
