+++ package.json
@@ -19,9 +19,6 @@
     "react-dom": "^18.3.1",
     "react-textarea-autosize": "^8.5.4",
     "react-toastify": "^9.0.1",
-    "slate": "^0.110.2",
-    "slate-history": "^0.110.3",
-    "slate-react": "^0.111.0",
     "tippy.js": "^6.3.7",
     "tiptap-markdown": "^0.8.10"
   },
