+++ public/pages/ShowPost/components/ShowComment.tsx
@@ -1,26 +1,13 @@
 import React, { useEffect, useRef, useState } from "react"
 import { Comment, Post, ImageUpload } from "@fider/models"
-import {
-  Reactions,
-  Avatar,
-  UserName,
-  Moment,
-  Form,
-  Button,
-  Markdown,
-  Modal,
-  ImageViewer,
-  MultiImageUploader,
-  Dropdown,
-  Icon,
-  CommentEditor,
-} from "@fider/components"
+import { Reactions, Avatar, UserName, Moment, Form, Button, Markdown, Modal, ImageViewer, MultiImageUploader, Dropdown, Icon } from "@fider/components"
 import { HStack } from "@fider/components/layout"
 import { formatDate, Failure, actions, notify, copyToClipboard, classSet, clearUrlHash } from "@fider/services"
 import { useFider } from "@fider/hooks"
 import IconDotsHorizontal from "@fider/assets/images/heroicons-dots-horizontal.svg"
 import { t } from "@lingui/core/macro"
 import { Trans } from "@lingui/react/macro"
+import CommentEditor from "@fider/components/common/form/CommentEditor"
 
 interface ShowCommentProps {
   post: Post
